Move localhost / loopbackinterface detection to TestUtils
